idLib to bot
