+++ Main.java
@@ -15,7 +15,7 @@ public class Main {
 
     public static int work(Scanner scan) {
         String command = bot.takeCommand(scan);
-        int idCommand = bot.commandToId(command);
+        int idCommand = idLib.getId(command);
         if (idCommand == -1) return 0;
         bot.quote(idCommand);
         return 1;
@@ -25,46 +25,5 @@ public class Main {
         System.out.print("Здарова! Введи какую-нить команду, например, /help \n");
     }
 
-    public static class bot {
-        public static String takeCommand(Scanner scan) {
-            String com = scan.nextLine();
-            return com;
-        }
 
-        public static int commandToId(String command) {
-            switch (command) {
-                case "exit":
-                    System.out.println("Пока-пока!");
-                    return -1;
-                case "help":
-                    return 7;
-                case "javascript":
-                    return 1000;
-                default:
-                    System.out.println("Я же сказал хелп введи, если не шаришь");
-                    return 0;
-            }
-        }
-
-        public static void quote(int id) {
-            try(BufferedReader br = new BufferedReader(new FileReader("quotes.txt")))
-            {
-                String s;
-                while((s=br.readLine())!=null){
-                    if (s.equals('$' + String.valueOf(id))) {
-                        break;
-                    }
-                }
-                while((s=br.readLine())!=null){
-                    if (s.equals("$")) {
-                        break;
-                    }
-                    System.out.println(s);
-                }
-            }
-            catch(IOException ex){
-                System.out.println(ex.getMessage());
-            }
-        }
-    }
 }
