+++ quotes.txt
@@ -6,7 +6,8 @@ $Ya xz lol$
 
 hhtr
 $7
-dg
+я тебе помогу
+gtlbr
 $
 hdfg
